Refactor: extract record operations from client.go into data.go
